Move testUtils to folder
