+++ apps/api/tests/api/collaborators-controller.test.ts
@@ -38,7 +38,7 @@ import {
 	PG_PASSWORD,
 	PG_USER,
 	testUserId as user_id,
-} from '../testUtils.js';
+} from '../utils/testUtils.ts';
 
 describe('Collaborators Controller', () => {
 	let db: PostgresDb;
@@ -196,7 +196,6 @@ describe('Collaborators Controller', () => {
 	});
 
 	after(async () => {
-		await db.delete(collaborators).where(eq(collaborators.application_id, application_id));
 		await container.stop();
 		process.exit(0);
 	});
